[UNI-172] chore: resolver 패키지 common 하위로 이동
