+++ uniro_backend/src/main/java/com/softeer5/uniro_backend/common/resolver/DangerListConverter.java
@@ -1,4 +1,4 @@
-package com.softeer5.uniro_backend.resolver;
+package com.softeer5.uniro_backend.common.resolver;
 
 import java.io.IOException;
 import java.util.Set;
@@ -7,18 +7,16 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.softeer5.uniro_backend.route.entity.CautionFactor;
+import com.softeer5.uniro_backend.route.entity.DangerFactor;
 
 import jakarta.persistence.AttributeConverter;
-import jakarta.persistence.Converter;
 
-@Converter
-public class CautionListConverter implements AttributeConverter<Set<CautionFactor>, String> {
+public class DangerListConverter implements AttributeConverter<Set<DangerFactor>, String>  {
 	private static final ObjectMapper mapper = new ObjectMapper()
 		.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false)
 		.configure(DeserializationFeature.FAIL_ON_NULL_FOR_PRIMITIVES, false);
 	@Override
-	public String convertToDatabaseColumn(Set<CautionFactor> attribute) {
+	public String convertToDatabaseColumn(Set<DangerFactor> attribute) {
 		try {
 			if (attribute == null) {
 				return "[]"; // List가 null일 경우, DB에 저장할 값은 []
@@ -30,7 +28,7 @@ public class CautionListConverter implements AttributeConverter<Set<CautionFacto
 	}
 
 	@Override
-	public Set<CautionFactor> convertToEntityAttribute(String dbData) {
+	public Set<DangerFactor> convertToEntityAttribute(String dbData) {
 		try {
 			if (dbData == null || dbData.trim().isEmpty()) {
 				return null; // dbData가 null 또는 빈 문자열일 경우 빈 리스트 반환
